+++ include/sdf/Element.hh
@@ -17,7 +17,9 @@
 #ifndef _SDF_ELEMENT_HH_
 #define _SDF_ELEMENT_HH_
 
+#include <map>
 #include <memory>
+#include <set>
 #include <string>
 #include <utility>
 #include <vector>
@@ -286,6 +288,29 @@ namespace sdf
     /// tree. First call parent->GetFirstElement() to get the first child. Call
     /// child = child->GetNextElement() to iterate through the children.
     public: ElementPtr GetNextElement(const std::string &_name = "") const;
+
+    /// \brief Get set of child element type names.
+    /// \return A set of the names of the child elements.
+    public: std::set<std::string> GetElementTypeNames() const;
+
+    /// \brief Checks whether any child elements of the specified element type
+    /// have identical name attribute values and returns false if so.
+    /// \param[in] _type The type of Element to check. If empty, check names
+    /// of all child elements.
+    /// \return True if all child elements with name attributes of the
+    /// specified type have unique names, return false if there are duplicated
+    /// names. Also return true if no elements of the specified type are found.
+    public: bool HasUniqueChildNames(const std::string &_type = "") const;
+
+    /// \brief Count the number of child elements of the specified element type
+    /// that have the same name attribute value.
+    /// \param[in] _type The type of Element to check. If empty, count names
+    /// of all child elements.
+    /// \return Map from Element names to a count of how many times the name
+    /// occurs. The count should never be 0. If all 2nd values are 1, then
+    /// there are exclusively unique names.
+    public: std::map<std::string, std::size_t>
+            CountNamedElements(const std::string &_type = "") const;
 
     /// \brief Return a pointer to the child element with the provided name.
     ///